fix(containers): rename redux action import
re #393
